+++ src/blocks/close-button/edit.js
@@ -2,6 +2,7 @@
  * WordPress dependencies
  */
 import { useBlockProps } from '@wordpress/block-editor';
+import { Notice } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
 
 export default function Edit() {
@@ -12,8 +13,19 @@ export default function Edit() {
 	} );
 
 	return (
-		<button { ...blockProps }>
-			<svg aria-hidden="true" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" width="24" height="24" fill="none" stroke="currentColor" strokeWidth="3" strokeLinecap="round" strokeLinejoin="round"><line x1="18" y1="6" x2="6" y2="18" /><line x1="6" y1="6" x2="18" y2="18" /></svg>
-		</button>
+		<>
+			<Notice
+				status="warning"
+				isDismissible={ false }
+			>
+				{ __(
+					'This block is deprecated. Use a Modal Trigger block with "Close modal" action instead.',
+					'pikari-gutenberg-modals'
+				) }
+			</Notice>
+			<button { ...blockProps }>
+				<svg aria-hidden="true" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" width="24" height="24" fill="none" stroke="currentColor" strokeWidth="3" strokeLinecap="round" strokeLinejoin="round"><line x1="18" y1="6" x2="6" y2="18" /><line x1="6" y1="6" x2="18" y2="18" /></svg>
+			</button>
+		</>
 	);
 }